added resume functionality

+++ src/Downloader.java
@@ -16,12 +16,16 @@ public class Downloader extends Thread {
     InputStream inputStream = null;
     OutputStream outStream = null;
     try {
+      File targetFile = new File(this.part.filePath);
+      if(targetFile.exists()){
+        System.out.println(this.part.filePath + " Already Exists");
+        return;
+      }
       HttpURLConnection urlConnection = (HttpURLConnection) this.part.url.openConnection();
       urlConnection.setRequestProperty("Range", "bytes=" + this.part.startBytes + "-" + this.part.endBytes);
       urlConnection.connect();
-
       inputStream = urlConnection.getInputStream();
-      File targetFile = new File(this.part.filePath);
+
       outStream = new FileOutputStream(targetFile);
       byte[] buffer = new byte[inputStream.available()];
       inputStream.read(buffer);
